Implement SetRpm

+++ devices/device.h
@@ -61,7 +61,7 @@ public:
     virtual SerialCode Close() override;
     virtual SerialCode Read() override;
     virtual SerialCode Write() override;
-    SerialCode SetRpm(BYTE channel, int rpm);
+    SerialCode SetRpm(BYTE channel, float rpm);
     SerialCode SetCw(BYTE channel);
     SerialCode SetCcw(BYTE channel);
     SerialCode On(BYTE channel);
